refactor : interceptor - repository -> service
interceptor에서 일부 repository에 직접 접근하던 것을 service를 통하도록 변경

issue key : SB2-6

+++ simpleBoard02/server-b-user/src/main/java/kim/zhyun/serveruser/interceptor/DisconnectionInterceptor.java
@@ -4,8 +4,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import kim.zhyun.serveruser.entity.SessionUser;
+import kim.zhyun.serveruser.service.NicknameStorageService;
+import kim.zhyun.serveruser.service.SessionUserRedisService;
 import kim.zhyun.serveruser.service.impl.NicknameStorageServiceImpl;
-import kim.zhyun.serveruser.repository.SessionUserRedisRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -17,8 +18,8 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Component
 public class DisconnectionInterceptor implements HandlerInterceptor {
-    private final SessionUserRedisRepository sessionUserRedisRepository;
-    private final NicknameStorageServiceImpl nicknameStorageService;
+    private final SessionUserRedisService sessionUserRedisService;
+    private final NicknameStorageService nicknameStorageService;
     
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
@@ -28,17 +29,17 @@ public class DisconnectionInterceptor implements HandlerInterceptor {
         log.info("delete nickname cache - session id: {}", sessionId);
 
         // 1. session_id storage : nickname 값 조회
-        Optional<SessionUser> optionalSessionUser = sessionUserRedisRepository.findById(sessionId);
+        Optional<SessionUser> optionalSessionUser = sessionUserRedisService.findById(sessionId);
         if (optionalSessionUser.isPresent()) {
             
             // 2. nickname storage : 1에서 조회한 nickname 삭제
             String nickname = optionalSessionUser.get().getNickname();
-            if (nickname != null || !nickname.isBlank()) {
+            if (nickname != null && !nickname.isBlank()) {
                 nicknameStorageService.deleteNickname(nickname);
             }
             
             // 3. session_id storage : session_id 삭제
-            sessionUserRedisRepository.deleteById(sessionId);
+            sessionUserRedisService.deleteById(sessionId);
         }
         
         return true;
